# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TÖPP API",
    "description": "REST API интернет-магазина ковров; цены в целых рублях. /docs для интерактивного OpenAPI.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/carts": {
      "post": {
        "summary": "Create Cart",
        "operationId": "create_cart_api_v1_carts_post",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cart": {
      "get": {
        "summary": "Get Cart",
        "operationId": "get_cart_api_v1_cart_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cartBearer": []
          }
        ]
      }
    },
    "/api/v1/cart/items": {
      "post": {
        "summary": "Add Item",
        "operationId": "add_item_api_v1_cart_items_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemIn"
              },
              "example": {
                "variantId": "00000000-0000-4000-8000-000000000000",
                "quantity": 2
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cartBearer": []
          }
        ]
      }
    },
    "/api/v1/cart/items/{item_id}": {
      "patch": {
        "summary": "Update Item",
        "operationId": "update_item_api_v1_cart_items__item_id__patch",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cartBearer": []
          }
        ]
      },
      "delete": {
        "summary": "Delete Item",
        "operationId": "delete_item_api_v1_cart_items__item_id__delete",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cartBearer": []
          }
        ]
      }
    },
    "/api/v1/favorites": {
      "get": {
        "summary": "Favorites",
        "operationId": "favorites_api_v1_favorites_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cartBearer": []
          }
        ]
      }
    },
    "/api/v1/favorites/{product_id}": {
      "put": {
        "summary": "Favorite Add",
        "operationId": "favorite_add_api_v1_favorites__product_id__put",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cartBearer": []
          }
        ]
      },
      "delete": {
        "summary": "Favorite Delete",
        "operationId": "favorite_delete_api_v1_favorites__product_id__delete",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cartBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemIn": {
        "properties": {
          "variantId": {
            "type": "string",
            "pattern": "^[0-9a-fA-F-]{36}$",
            "title": "Variantid"
          },
          "quantity": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Quantity"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "variantId",
          "quantity"
        ],
        "title": "ItemIn"
      },
...
      "QuantityIn": {
        "properties": {
          "quantity": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Quantity"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "quantity"
        ],
        "title": "QuantityIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "cartBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Токен из POST /api/v1/carts"
      },
      "adminBasic": {
        "type": "http",
        "scheme": "basic",
        "description": "ADMIN_USERNAME и пароль администратора"
      }
    }
  }
}